# clinica_api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clínica Social Cuidar - API System",
    "description": "\n**API Oficial de Gestão da Clínica Social Cuidar.**\n\n## Funcionalidades Principais:\n*   **Gestão de Pacientes**: Cadastro completo com criptografia de dados sensíveis (LGPD).\n*   **Prontuário Eletrônico**: Registro histórico de atendimentos.\n*   **Controle Financeiro**: Receitas e Despesas.\n*   **Auditoria**: Rastreabilidade total de ações.\n\n## Segurança e Governança:\n*   Todos os endpoints são protegidos (JWT).\n*   Rate Limiting ativo para prevenção de ataques.\n*   Dados sensíveis (CPF) são criptografados no banco (Fernet).\n        ",
    "contact": {
      "name": "Suporte Técnico - Vitor Moraes",
      "email": "suporte@clinicasocial.com.br"
    },
    "license": {
      "name": "Private License"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/patients": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Listar todos os pacientes",
        "description": "Retorna uma lista de pacientes ativos. **Dados sensíveis (CPF) são descriptografados automaticamente** para visualização.",
        "operationId": "read_patients_api_v1_patients_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Patient"
                  },
                  "type": "array",
                  "title": "Response Read Patients Api V1 Patients Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "patients"
        ],
        "summary": "Cadastrar novo paciente",
        "description": "Cria um novo registro de paciente. **O CPF é criptografado automaticamente** antes de ser salvo no banco de dados.",
        "operationId": "create_patient_api_v1_patients_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientBase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/patients/{patient_id}": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Obter detalhes de um paciente",
        "description": "Busca um paciente pelo ID. Retorna 404 se não encontrado ou inativo.",
        "operationId": "read_patient_api_v1_patients__patient_id__get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "patients"
        ],
        "summary": "Atualizar dados do paciente",
        "description": "Atualiza campos do paciente. Se o CPF for alterado, ele será criptografado novamente.",
        "operationId": "update_patient_api_v1_patients__patient_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientBase"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "patients"
        ],
        "summary": "Arquivar paciente (Soft Delete)",
        "description": "Marca o paciente como inativo (`active=False`). Não remove do banco para manter integridade do histórico.",
        "operationId": "delete_patient_api_v1_patients__patient_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/{patient_id}/anonymize": {
      "post": {
        "tags": [
          "patients"
        ],
        "summary": "Anonimizar paciente (Direito ao Esquecimento)",
        "description": "**Ação Irreversível.** Substitui dados pessoais por valores anonimizados, mantendo apenas o ID e histórico clínico estatístico.",
        "operationId": "anonymize_patient_api_v1_patients__patient_id__anonymize_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/forms/templates": {
      "post": {
        "tags": [
          "forms"
        ],
        "summary": "Create Template",
        "operationId": "create_template_api_v1_forms_templates_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormTemplateCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormTemplateRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "forms"
        ],
        "summary": "Read Templates",
        "operationId": "read_templates_api_v1_forms_templates_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "specialty",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Specialty"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FormTemplateRead"
                  },
                  "title": "Response Read Templates Api V1 Forms Templates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forms/templates/{template_id}": {
      "delete": {
        "tags": [
          "forms"
        ],
        "summary": "Delete Template",
        "operationId": "delete_template_api_v1_forms_templates__template_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "forms"
        ],
        "summary": "Update Template",
        "operationId": "update_template_api_v1_forms_templates__template_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormTemplateCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormTemplateRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FormTemplateCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "dynamic"
          },
          "schema_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Schema Config",
            "default": {}
          },
          "specialties": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Specialties",
            "default": []
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "FormTemplateCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Patient": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Nome completo do paciente",
            "example": "João da Silva"
          },
          "cpf": {
            "type": "string",
            "title": "Cpf",
            "description": "CPF do paciente (será criptografado)",
            "example": "123.456.789-00"
          },
          "rg": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rg",
            "description": "RG do paciente",
            "example": "12.345.678-9"
          },
          "birth_date": {
            "type": "string",
            "title": "Birth Date",
            "description": "Data de nascimento (YYYY-MM-DD)",
            "example": "1980-01-01"
          },
          "whatsapp": {
            "type": "string",
            "title": "Whatsapp",
            "description": "Número de WhatsApp com DDD",
            "example": "11999999999"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "Email de contato",
            "example": "joao@email.com"
          },
          "address": {
            "$ref": "#/components/schemas/Address",
            "description": "Endereço completo"
          },
          "personal_income": {
            "type": "number",
            "title": "Personal Income",
            "description": "Renda pessoal mensal",
            "example": 1500.0
          },
          "family_income": {
            "type": "number",
            "title": "Family Income",
            "description": "Renda familiar mensal",
            "example": 3000.0
          },
          "observations": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observations",
            "description": "Observações gerais",
            "example": "Paciente diabético"
          },
          "files": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Files",
            "description": "Lista de arquivos anexados (URLs)",
            "default": []
          },
          "photo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo",
            "description": "URL da foto do perfil"
          },
          "payment_table_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Table Id",
            "description": "ID da tabela de preços associada"
          },
          "guardian_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Name",
            "description": "Nome do responsável",
            "example": "Maria da Silva"
          },
          "guardian_cpf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Cpf",
            "description": "CPF do responsável",
            "example": "987.654.321-00"
          },
          "guardian_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Phone",
            "description": "Telefone do responsável",
            "example": "11988888888"
          },
          "lgpd_consent": {
            "type": "boolean",
            "title": "Lgpd Consent",
            "description": "Consentimento LGPD assinado?",
            "default": false
          },
          "lgpd_consent_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lgpd Consent Date",
            "description": "Data do consentimento (ISO)"
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "ID único do sistema (UUID)"
          }
        },
        "type": "object",
        "required": [
          "name",
          "cpf",
          "birth_date",
          "whatsapp",
          "address",
          "personal_income",
          "family_income",
          "id"
        ],
        "title": "Patient"
      },
      "PatientBase": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Nome completo do paciente",
            "example": "João da Silva"
          },
          "cpf": {
            "type": "string",
            "title": "Cpf",
            "description": "CPF do paciente (será criptografado)",
            "example": "123.456.789-00"
          },
          "rg": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rg",
            "description": "RG do paciente",
            "example": "12.345.678-9"
          },
          "birth_date": {
            "type": "string",
            "title": "Birth Date",
            "description": "Data de nascimento (YYYY-MM-DD)",
            "example": "1980-01-01"
          },
          "whatsapp": {
            "type": "string",
            "title": "Whatsapp",
            "description": "Número de WhatsApp com DDD",
            "example": "11999999999"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "Email de contato",
            "example": "joao@email.com"
          },
          "address": {
            "$ref": "#/components/schemas/Address",
            "description": "Endereço completo"
          },
          "personal_income": {
            "type": "number",
            "title": "Personal Income",
            "description": "Renda pessoal mensal",
            "example": 1500.0
          },
          "family_income": {
            "type": "number",
            "title": "Family Income",
            "description": "Renda familiar mensal",
            "example": 3000.0
          },
          "observations": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observations",
            "description": "Observações gerais",
            "example": "Paciente diabético"
          },
          "files": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Files",
            "description": "Lista de arquivos anexados (URLs)",
            "default": []
          },
          "photo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo",
            "description": "URL da foto do perfil"
          },
          "payment_table_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Table Id",
            "description": "ID da tabela de preços associada"
          },
          "guardian_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Name",
            "description": "Nome do responsável",
            "example": "Maria da Silva"
          },
          "guardian_cpf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Cpf",
            "description": "CPF do responsável",
            "example": "987.654.321-00"
          },
          "guardian_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Phone",
            "description": "Telefone do responsável",
            "example": "11988888888"
          },
          "lgpd_consent": {
            "type": "boolean",
            "title": "Lgpd Consent",
            "description": "Consentimento LGPD assinado?",
            "default": false
          },
          "lgpd_consent_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lgpd Consent Date",
            "description": "Data do consentimento (ISO)"
          }
        },
        "type": "object",
        "required": [
          "name",
          "cpf",
          "birth_date",
          "whatsapp",
          "address",
          "personal_income",
          "family_income"
        ],
        "title": "PatientBase"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}